Introduce API resource for creating a fileset
This adds a /filesets resource and a POST handler to create a fileset
from an incoming source bundle. The rough design is:

- A source bundle (file or files) to generate a fileset are posted,
  along with a notional "target collection", "name", and "profile"
- The fileset identifer combines the collection and name, hashing it and
  converting to UUID format
- The identifier serves as a base working directory for that fileset
- An ordinal "job" directory is created inside to handle re-runs (for
  whatever reason) where the identifier would be retained
- The uploaded files are copied to a "src" directory for the job
- The assembled fileset should be written to the job's "build" directory

There are many assumptions and prototype details here; it is not to be
taken as final form, but to sketch the ideas of transfer via HTTP,
identifying and isolating filesets/job workspaces, and selecting a
profile for the file/fileset.