Add File Set job and worker
This is further prototyping with job queueing and processing, using
Redis and RQ. It splits the previous step into two halves: setting up
the work directory (API) and the "actual" processing (job/worker).

Because there is now shared storage required beneath multiple
containers, there is a bit of messing around at the Compose file
required. This uses a named volume (rather than mapping the local
source directory, to avoid pollution and simplify cleanup). However,
this requires ensuring that the permissions are set for the user that
the API and worker use. A one-shot sidecar container takes care of that
as a startup dependency.

We mount the fileset working storage at /data/filesets, conventionally.
This can probably be consolidated with the workspaces, but I did not
want to collide with the other setup and expectations while prototyping.

The actual "job" (the process_fileset function) is moved to a standalone
file that can be imported, and the worker script can call call it. All
of that setup is a bit sloppy and changeable. The relationship between
app, jobs, and workers is typically very blurry in Resque-inspired
tools, and there are plenty of options for modeling.

As it stands, only one "profile" (basic-image) is simulated. There is
currently no distinction between the generic invocation of file set
processing and the specific modality of a profile. These profiles are
intended to be analogous to "workflows", and there may be steps or
transforms that are used between them, and they will probably all fall
under a general/high-level shape of "file set processing".

To see this in action:

- `docker compose up --build`
- Visit http://127.0.0.1:8000/docs#/default/create_fileset_api_v1_filesets__post
- Submit a file (or multiple) with the basic-image profile
- Check for logs at the API and worker
- `docker compose exec api ls -laR /data/filesets`