Improve fileset process organization somewhat
There are still some outstanding questions about naming and organization
of the pieces of this code, but in the interest of merging and then
discussing improvements, they are merely annotated with comments for the
time being.

This change adds a `bin` directory for "application scripts", but does
not split the pieces of file set processing that are needed from the API
and the generic worker -- they are combined as a grab bag under
providers.