Update Typer and get subcommands ready for package
There is a subtle dependency problem with click 8.2 that causes the
packaged binary to fail. Typer 0.15.4 was released to limit that
dependency range and resolve it. There are no other package changes
here.

We also set the help flag by default since the top-level and two
sub-commands in place have no default behavior. Because the app will be
packaged rather than run with python <whatever.py>, we also add the
callback decorator and remove the __main__ check. This is replaced by
the `dor` binary that poetry installs.

See also:

- https://typer.tiangolo.com/tutorial/package/#create-your-app
- https://github.com/fastapi/typer/pull/1225