Fix up loading of default text font
The sample generator was originally loading from a simple relative path.
This changes it to load a bundled resource. There are four key pieces:

- Include all three packages (dor, gateway, utils)
- Include the etc directory (for both sdist and wheel)
- Use importlib to locate and load the font file resource
- Move the `text_font` to Settings

The design here introduces a small wrapper dataclass for specifying a
font file within the root fonts directory. All it really does is
separate the pieces of the path/filename into identified components and
then assemble the full path from there. There is a deferred `load()`
function to create the FreeTypeFont object.

This is currently invoked upon startup and stored on the Settings
object. For now, it only changes the reference from the file_set sample
generator (from settings.text_font to settings.S.text_font). We may
decide that Settings should only be primitives rather than loaded
resources, but I wanted to eliminate the top-level variable for the sake
of discussion. The scope of that change is small and easy to revert.

As a trivial change, the dist/ directory is added to .gitignore to avoid
accidental commits of assembled packages.

See also:

- https://python-poetry.org/docs/pyproject/#exclude-and-include
- https://docs.python.org/3/library/importlib.resources.html#module-importlib.resources