Switch from build args to build secrets
The build detected the use of the password as an ARG and warned against
it, so we switch to mounted secrets. It's a little more verbose on the
Dockerfile side, but more in line with recommendations. There is also a
`secret-envs` option for build-push-action that may work, but the exact
behavior is unclear (possibly taking the value _from_ an env var).