Declare Dockerfile syntax and context ignores
The exact syntax to support secret mounts as environment variables could
cause problems with a syntax level declaration. Adding the magic comment
with syntax=docker/dockerfile:1 uses the latest v1 syntax.

We also ignore .venv/ and .git/ directories in the build context, which
reduces it by at least 100MB. If using a development mount of the
working directory, these will still be available -- this change only
avoids copying the files into the isolated build context, where they
should probably never be used. It is unlikely that a host venv is
correct for operating within the container. It is also unlikely that a
copy of the git repository in the container is of use. We definitely do
not want to include them in images.

See also: https://github.com/docker/docs/issues/20935#issuecomment-2363234867