Pass UoW to event handlers and test registration
The message bus was not passing the unit of work along to the handler
upon invocation, so the tests were set up to bind it into the handler
with a closure. Here, we fix that and test the dynamic registration.

There will be follow-on work in framework.py, but different handlers
will need other dependencies injected, so it's not as direct a change.