Normalize dependency injection for messagebus/uow
We now use handlers with all of their dependencies bound, so the message
bus can invoke them with only the message. To do so, we also inject the
unit of work instance into the message bus. This is primarily because we
use the uow to collect events.

We also add a bit of test support code for spying values of events and
ensure that commands/events can emit other events that are eventually
handled.