Rough in external messaging with RabbitMQ/Pika
- Add CreatePackage command
- Add rabbitmq compose service
- Add basic packaging worker/consumer and compose service
- Add a basic message publisher (as module/function)
- Switch /packages POST to emitting CreatePackage command
- Switch create_package_from_metadata to emitting PackageSubmitted event
  rather than enqueuing automation job

There are quite a few items that we still need to consider and decide,
but this gets the fundamentals in place for messaging between
services/subdomains.

The publisher and subscriber follow the simple example from Architecture
Patterns with Python, though we will want to consider a broker object
with a single connection and possibly multiple channels. All of the
exchange type, routing, message properties, and mapping of events to
queues/topics or handlers is up for debate.

We also need to consider the separation and semantics of the external
vs. internal messages, which should inform our domain boundaries and
support our reorganization of the source modules along those lines.