Integrate packaging and ingest with events
- Add ingest-worker
- Emit PackageGenerated when completing the package
- Consume PackageGenerated in the ingest-worker
- Emit PackageIngested upon faked success of ingest

This is definitely not done; there are a couple of semantic issues as
well as organizational ones. The biggest problem is that the packaging
and ingest workers currently compete on the same work queue.

The path forward should be to distinguish between the queue for
commands, which will be consumed within a subdomain, and the topic for
events, which may be consumed by any number of listeners (inside or
outside of the subdomain).